requirements.txt
fastapi==0.111.0
python-multipart==0.0.9
pdfminer.six==20231228
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
